requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
weasyprint>=60.0
google-genai
google-adk
firebase-admin>=6.0.0
markdown>=3.4.0
requests>=2.31.0 